requirements.txt
Django==4.2.2
psycopg2-binary==2.9.6
Pillow==9.5.0
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
djangorestframework==3.14.0
pyairports==2.1.1


